requirements.txt
fastapi==0.109.2
langchain==0.1.5
langchain-community==0.0.19
langchain-core==0.1.21
langchain-openai==0.0.5
pydantic==2.6.1
PyPDF2==3.0.1
python-dotenv==1.0.1
python-multipart==0.0.7
uvicorn==0.27.0.post1
streamlit
chromadb
ibm-watson
urllib3
ibm-watson-machine-learning
huggingface-hub
sentence-transformers
tiktoken